refactor: BaseEntity 제거, 엔티티에 Auditing 개별 설정
