--- src/main/java/kitchenpos/BaseEntity.java
@@ -1,19 +1,0 @@
-package kitchenpos;
-
-import java.time.LocalDateTime;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-
-    @CreatedDate
-    private LocalDateTime createdDate;
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-}
