+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,22 +1,33 @@
 package kitchenpos.tablegroup.domain;
 
+import java.time.LocalDateTime;
 import javax.persistence.Entity;
+import javax.persistence.EntityListeners;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.BaseEntity;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
-public class TableGroup extends BaseEntity {
+@EntityListeners(AuditingEntityListener.class)
+public class TableGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @CreatedDate
+    private LocalDateTime createdDate;
 
     public TableGroup() {
     }
 
     public Long getId() {
         return id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
     }
 }
