+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.math.BigDecimal;
-import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -184,6 +183,6 @@ class TableGroupServiceTest extends ServiceTest {
         OrderLineItem orderItem1 = new OrderLineItem(menu.getId(), 1L);
         OrderLineItem orderItem2 = new OrderLineItem(menu.getId(), 1L);
         final List<OrderLineItem> orderLineItems = List.of(orderItem1, orderItem2);
-        return new Order(orderTable.getId(), LocalDateTime.now(), orderLineItems);
+        return new Order(orderTable.getId(), orderLineItems);
     }
 }
